refactor(journal): split service.go below 1000 lines
Move helpers into sibling files in package journal.
No behavior change.

Related: #652
